create a separate getReactDOMServer that handles importing react-dom/server.edge
