+++ packages/render/src/shared/read-stream.ts
